refactor: 불필요한 코드 수정 및 주석 추가;
