+++ src/hooks/useApplyScreenHooks.ts
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
 import {useEffect, useState} from 'react';
-import {TodoTestSchema} from './realm/TodoTestSchema';
+import {TodoTestSchema} from '../realm/TodoTestSchema';
 import {Alert} from 'react-native';
 import Realm from 'realm';
 
@@ -25,11 +25,13 @@ const useApplyScreenHooks = () => {
   //   });
   // };
 
+  // * 입력후 전달하는 sumit function
   const onSubmit = () => {
     if (inputTodo.length > 0) {
-      setInputTodo('');
-      // realm todo create
+      // DB에 입력값 전달
       createDB(inputTodo);
+      // 입력 후 input init
+      setInputTodo('');
     } else {
       Alert.alert('알림', '할일을 적어주세요!', [
         {text: '확인', onPress: () => {}},
@@ -42,14 +44,18 @@ const useApplyScreenHooks = () => {
   // ==================
   // * relam config
   // The initial schemaVersion is 0.
+
   const realm = new Realm({
     schema: [TodoTestSchema],
   });
+
+  // *TodoList 데이터 Read
   const readTodo = () => {
     const result = realm.objects('todotest').toJSON() as TodoListType[];
     setTodoList(result);
-    // console.log(result);
   };
+
+  // *Todolist localDb에 등록
   const createDB = (title: string) => {
     const id = new Realm.BSON.UUID();
     realm.write(() => {
@@ -60,29 +66,31 @@ const useApplyScreenHooks = () => {
         create_at: dayjs().toString(),
       });
     });
+
     readTodo();
   };
+
+  // 데이터 삭제 후 갱신
   const deleteTodoList = (id: string) => {
     const deleteDB = realm.objects('todotest').filtered(`id = '${id}'`)[0];
     realm.write(() => {
       realm.delete(deleteDB);
     });
 
-    // 삭제 후 데이터 갱신
     readTodo();
   };
+
+  // 데이터 상태 업데이트 후 갱신
   const updateTodoList = (id: string) => {
     const targetData = realm.objects('todotest').filtered(`id = '${id}'`)[0];
-    // console.log('[targetData] : ', targetData);
-
     realm.write(() => {
       targetData.isChecked = !targetData.isChecked;
     });
 
-    // 수정 후 데이터 갱신
     readTodo();
   };
 
+  // 최초 localDB Read
   useEffect(() => {
     readTodo();
     // eslint-disable-next-line react-hooks/exhaustive-deps
